Relayout `constraints` module
- `api` base and "meta" classes
- `basic` constraint primitives for basic data types
- `utils` some (internal) helpers (at least for now)

This also removed some functionality

- `NoneDeprecated`: "A helper construct to ease migrations for
  option=None"

  Introduced with 3c4187ff268b70377dbb91ee0b6ce725ce74a630 more than
  a year ago. It is only used in `datalad/local/subdatasets.py`.

- `expand_constraint_spec()`: Can only handle `float`, `int`, `bool`,
  `str`, and is only used in `Parameter`.
