+++ datalad_next/constraints/tests/test_constraints.py
@@ -1,24 +1,31 @@
-# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
-# vi: set ft=python sts=4 ts=4 sw=4 et:
-### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-#
-#   See COPYING file distributed along with the DataLad package for the
-#   copyright and license terms.
-#
-### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-'''Unit tests for basic constraints functionality.'''
-
-
 from datalad.tests.utils_pytest import (
     assert_equal,
     assert_raises,
 )
 
-from ... import constraints as ct
+
+from ..api import (
+    Constraints,
+    AltConstraints,
+)
+from ..basic import (
+    EnsureInt,
+    EnsureFloat,
+    EnsureBool,
+    EnsureStr,
+    EnsureNone,
+    EnsureCallable,
+    EnsureChoice,
+    EnsureKeyChoice,
+    EnsureRange,
+    EnsureListOf,
+    EnsureTupleOf,
+)
+from ..utils import _type_str
 
 
 def test_int():
-    c = ct.EnsureInt()
+    c = EnsureInt()
     # this should always work
     assert_equal(c(7), 7)
     assert_equal(c(7.0), 7)
@@ -33,7 +40,7 @@ def test_int():
 
 
 def test_float():
-    c = ct.EnsureFloat()
+    c = EnsureFloat()
     # this should always work
     assert_equal(c(7.0), 7.0)
     assert_equal(c(7), 7.0)
@@ -45,7 +52,7 @@ def test_float():
 
 
 def test_bool():
-    c = ct.EnsureBool()
+    c = EnsureBool()
     # this should always work
     assert_equal(c(True), True)
     assert_equal(c(False), False)
@@ -69,7 +76,7 @@ def test_bool():
 
 
 def test_str():
-    c = ct.EnsureStr()
+    c = EnsureStr()
     # this should always work
     assert_equal(c('hello'), 'hello')
     assert_equal(c('7.0'), '7.0')
@@ -82,36 +89,34 @@ def test_str():
     assert_equal(c.short_description(), 'str')
 
 def test_str_min_len():
-    c = ct.EnsureStr(min_len=1)
+    c = EnsureStr(min_len=1)
     assert_equal(c('hello'), 'hello')
     assert_equal(c('h'), 'h')
     assert_raises(ValueError, c, '')
 
-    c = ct.EnsureStr(min_len=2)
+    c = EnsureStr(min_len=2)
     assert_equal(c('hello'), 'hello')
     assert_raises(ValueError, c, 'h')
 
 
 def test_none():
-    c = ct.EnsureNone()
+    c = EnsureNone()
     # this should always work
     assert_equal(c(None), None)
-    # instance of NoneDeprecated is also None
-    assert_equal(c(ct.NoneDeprecated), None)
     # this should always fail
     assert_raises(ValueError, lambda: c('None'))
     assert_raises(ValueError, lambda: c([]))
 
 
 def test_callable():
-    c = ct.EnsureCallable()
+    c = EnsureCallable()
     # this should always work
     assert_equal(c(range), range)
     assert_raises(ValueError, c, 'range')
 
 
 def test_choice():
-    c = ct.EnsureChoice('choice1', 'choice2', None)
+    c = EnsureChoice('choice1', 'choice2', None)
     # this should always work
     assert_equal(c('choice1'), 'choice1')
     assert_equal(c(None), None)
@@ -121,7 +126,7 @@ def test_choice():
 
 
 def test_keychoice():
-    c = ct.EnsureKeyChoice(key='some', values=('choice1', 'choice2', None))
+    c = EnsureKeyChoice(key='some', values=('choice1', 'choice2', None))
     assert_equal(c({'some': 'choice1'}), {'some': 'choice1'})
     assert_equal(c({'some': None}), {'some': None})
     assert_equal(c({'some': None, 'ign': 'ore'}), {'some': None, 'ign': 'ore'})
@@ -133,7 +138,7 @@ def test_keychoice():
 
 
 def test_range():
-    c = ct.EnsureRange(min=3, max=7)
+    c = EnsureRange(min=3, max=7)
     # this should always work
     assert_equal(c(3.0), 3.0)
 
@@ -146,7 +151,7 @@ def test_range():
     assert_raises(TypeError, lambda: c('7'))
 
     # Range doesn't have to be numeric
-    c = ct.EnsureRange(min="e", max="qqq")
+    c = EnsureRange(min="e", max="qqq")
     assert_equal(c('e'), 'e')
     assert_equal(c('fa'), 'fa')
     assert_equal(c('qq'), 'qq')
@@ -155,59 +160,59 @@ def test_range():
 
 
 def test_listof():
-    c = ct.EnsureListOf(str)
+    c = EnsureListOf(str)
     assert_equal(c(['a', 'b']), ['a', 'b'])
     assert_equal(c(['a1', 'b2']), ['a1', 'b2'])
 
 
 def test_tupleof():
-    c = ct.EnsureTupleOf(str)
+    c = EnsureTupleOf(str)
     assert_equal(c(('a', 'b')), ('a', 'b'))
     assert_equal(c(('a1', 'b2')), ('a1', 'b2'))
 
 
 def test_constraints():
     # this should always work
-    c = ct.Constraints(ct.EnsureFloat())
+    c = Constraints(EnsureFloat())
     assert_equal(c(7.0), 7.0)
-    c = ct.Constraints(ct.EnsureFloat(), ct.EnsureRange(min=4.0))
+    c = Constraints(EnsureFloat(), EnsureRange(min=4.0))
     assert_equal(c(7.0), 7.0)
     # __and__ form
-    c = ct.EnsureFloat() & ct.EnsureRange(min=4.0)
+    c = EnsureFloat() & EnsureRange(min=4.0)
     assert_equal(c(7.0), 7.0)
     assert_raises(ValueError, c, 3.9)
-    c = ct.Constraints(ct.EnsureFloat(), ct.EnsureRange(min=4), ct.EnsureRange(max=9))
+    c = Constraints(EnsureFloat(), EnsureRange(min=4), EnsureRange(max=9))
     assert_equal(c(7.0), 7.0)
     assert_raises(ValueError, c, 3.9)
     assert_raises(ValueError, c, 9.01)
     # __and__ form
-    c = ct.EnsureFloat() & ct.EnsureRange(min=4) & ct.EnsureRange(max=9)
+    c = EnsureFloat() & EnsureRange(min=4) & EnsureRange(max=9)
     assert_equal(c(7.0), 7.0)
     assert_raises(ValueError, c, 3.99)
     assert_raises(ValueError, c, 9.01)
     # and reordering should not have any effect
-    c = ct.Constraints(ct.EnsureRange(max=4), ct.EnsureRange(min=9), ct.EnsureFloat())
+    c = Constraints(EnsureRange(max=4), EnsureRange(min=9), EnsureFloat())
     assert_raises(ValueError, c, 3.99)
     assert_raises(ValueError, c, 9.01)
 
 
 def test_altconstraints():
     # this should always work
-    c = ct.AltConstraints(ct.EnsureFloat())
+    c = AltConstraints(EnsureFloat())
     assert_equal(c(7.0), 7.0)
-    c = ct.AltConstraints(ct.EnsureFloat(), ct.EnsureNone())
+    c = AltConstraints(EnsureFloat(), EnsureNone())
     assert_equal(c.short_description(), '(float or None)')
     assert_equal(c(7.0), 7.0)
     assert_equal(c(None), None)
     # __or__ form
-    c = ct.EnsureFloat() | ct.EnsureNone()
+    c = EnsureFloat() | EnsureNone()
     assert_equal(c(7.0), 7.0)
     assert_equal(c(None), None)
 
     # this should always fail
-    c = ct.Constraints(ct.EnsureRange(min=0, max=4), ct.EnsureRange(min=9, max=11))
+    c = Constraints(EnsureRange(min=0, max=4), EnsureRange(min=9, max=11))
     assert_raises(ValueError, c, 7.0)
-    c = ct.EnsureRange(min=0, max=4) | ct.EnsureRange(min=9, max=11)
+    c = EnsureRange(min=0, max=4) | EnsureRange(min=9, max=11)
     assert_equal(c(3.0), 3.0)
     assert_equal(c(9.0), 9.0)
     assert_raises(ValueError, c, 7.0)
@@ -216,16 +221,17 @@ def test_altconstraints():
 
 def test_both():
     # this should always work
-    c = ct.AltConstraints(
-        ct.Constraints(
-            ct.EnsureFloat(),
-            ct.EnsureRange(min=7.0, max=44.0)),
-        ct.EnsureNone())
+    c = AltConstraints(
+        Constraints(
+            EnsureFloat(),
+            EnsureRange(min=7.0, max=44.0)),
+        EnsureNone())
     assert_equal(c(7.0), 7.0)
     assert_equal(c(None), None)
     # this should always fail
     assert_raises(ValueError, lambda: c(77.0))
 
+
 def test_type_str():
-    assert_equal(ct._type_str((str,)), 'str')
-    assert_equal(ct._type_str(str), 'str')
+    assert_equal(_type_str((str,)), 'str')
+    assert_equal(_type_str(str), 'str')
